Add DuelController class and calculateDuelScenario algorithm
DuelController class is used to manage boxers according to the scenario.

calculateDuelScenario function returns the duel scenario
based on boxers' physique and random
and must be moved to the backend.
Now it generates timings and names of moves for both left and right boxers.

Add files of constants duelParameters and duelAnimationsNames.