Hide model on switchLeadingSide bug
The bug suddenly appears when calling switchLeadingSide
under various unclear circumstances.
It seems that we need to inverse angle in face function
in some cases. Maybe not.

If we use lookAt from THREE.js, the bug with
model rotation disappears, but the old bug with
camera rotation appears instead.

It seems these problems are same, but are manifested
in different places.

So the bug has been hidden by setting the chance of
switchLeadingSide movement to zero.
Further investigation and fixing required.

Apply style and documentation fixes.